divided utils decoder into decoderMessage and decoderTx
