+++ src/utils/index.ts
@@ -1,4 +1,6 @@
 export * from './icons';
 export * from './utils';
 export * from './address';
-export * from './decoder';
+export * from './decoder/decoderUtils';
+export * from './decoder/decoderMessage';
+export * from './decoder/decoderTx';
